// bms-next/src/app/(protected)/(apps)/sales-crm/components/type.ts
export interface DealData {
  dealIdentifier: string;
  dealNo?: string;
  dealName?: string;
  expectedRevenue?: number;
  currency?: string;
  dealStatus?: string;
  productDetails: {
    [key: string]: {
      description: string;
      productIdentifier: string
      productType: string
      productDescription: string
      projectManager: string
      discountPercent?: number;
    }
  }
  checkedproductIdentifierList?: string[];
  updatedOn?: string;
  accountDetails?: {
    accountIdentifier: string;
    isChannelPartner?: boolean;
    wasFromChannelPartner?: boolean;
    accountType?: string;
    accountName?: string;
    accountManager?: string;
    accountCode?: string;
    taxinfo?: string;
    address?: string;
    city?: string;
    state?: string;
    pinCode?: string;
    country?: string;
    phone?: string;
    email?: string;
  };
  isDebtRevenue?: boolean;
  copyDealCheck?: boolean;
  olddealIdentifier?: string;
  dealStartDate?: string;
  createdBy?: string;
  vatPercent?: number;
  discountPercent?: number;
  fxRateForDeal?: Record<string, {
    id: string;
    currency: string;
    fxRate: number;
    year: string;
  }>;
  contractNumber?: string;
  dealWonDate?: string;
  formattedActualRevenueIncludingVAT_contracted?: string;
  contractedStartDate?: string;
  contractedEndDate?: string;
  projectName?: string;
  startDate?: string;
  parentProjectNo?: string;
  projectManager?: string;
  issueLinked?: boolean;
  dealTeam?: any[];
  projectManagerDelegates?: string;
  projectTeamUnderProjectManager?: any[];
  projectTeamUnderProjectManagerDelegates?: any[];
  activeStatus?: string;
  remarkForDealLostOrSuspended?: string;
  inActiveStatusDate?: string;
  accountName?: string;
  clientName?: string;
  dealProductTypeArray?: string[];
  source?: string;
  dealClosingDate?: string;
  shortProductName?: string;
  leadIdentifier?: string;
}